refactor(plugin-file-manager): adjust getFileData API
