+++ packages/plugins/@nocobase/plugin-file-manager/src/common/constants.ts
@@ -1,0 +1,10 @@
+/**
+ * This file is part of the NocoBase (R) project.
+ * Copyright (c) 2020-2024 NocoBase Co., Ltd.
+ * Authors: NocoBase Team.
+ *
+ * This project is dual-licensed under AGPL-3.0 and NocoBase Commercial License.
+ * For more information, please refer to: https://www.nocobase.com/agreement.
+ */
+
+export const INVALID_FILENAME_CHARS = '<>?*~\\/';
