+++ packages/plugins/@nocobase/plugin-file-manager/src/server/server.ts
@@ -7,18 +7,16 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
+import { basename } from 'path';
+import fs from 'fs';
+
 import { Plugin } from '@nocobase/server';
 import { isURL, Registry } from '@nocobase/utils';
-
-import { basename } from 'path';
-
 import { Collection, Model, Transactionable } from '@nocobase/database';
-import fs from 'fs';
 import { STORAGE_TYPE_ALI_OSS, STORAGE_TYPE_LOCAL, STORAGE_TYPE_S3, STORAGE_TYPE_TX_COS } from '../constants';
 import initActions from './actions';
-import { getFileData } from './actions/attachments';
 import { AttachmentInterface } from './interfaces/attachment-interface';
-import { AttachmentModel, StorageClassType, StorageModel, StorageType } from './storages';
+import { AttachmentModel, StorageClassType, StorageModel } from './storages';
 import StorageTypeAliOss from './storages/ali-oss';
 import StorageTypeLocal from './storages/local';
 import StorageTypeS3 from './storages/s3';
@@ -106,39 +104,28 @@ export class PluginFileManagerServer extends Plugin {
 
   async uploadFile(options: UploadFileOptions) {
     const { storageName, filePath, documentRoot } = options;
-    const storageRepository = this.db.getRepository('storages');
-    let storageInstance;
 
-    storageInstance = await storageRepository.findOne({
-      filter: storageName
-        ? {
-            name: storageName,
-          }
-        : {
-            default: true,
-          },
-    });
+    const storages = Array.from(this.storagesCache.values());
+    const storage = storages.find((item) => item.name === storageName) || storages.find((item) => item.default);
 
-    const fileStream = fs.createReadStream(filePath);
-
-    if (!storageInstance) {
+    if (!storage) {
       throw new Error('[file-manager] no linked or default storage provided');
     }
 
-    storageInstance = this.parseStorage(storageInstance);
+    const fileStream = fs.createReadStream(filePath);
 
     if (documentRoot) {
-      storageInstance.options['documentRoot'] = documentRoot;
+      storage.options['documentRoot'] = documentRoot;
     }
 
-    const storageType = this.storageTypes.get(storageInstance.type);
-    const storage = new storageType(storageInstance);
+    const StorageType = this.storageTypes.get(storage.type);
+    const storageInstance = new StorageType(storage);
 
-    if (!storage) {
-      throw new Error(`[file-manager] storage type "${storageInstance.type}" is not defined`);
+    if (!storageInstance) {
+      throw new Error(`[file-manager] storage type "${storage.type}" is not defined`);
     }
 
-    const engine = storage.make();
+    const engine = storageInstance.make();
 
     const file = {
       originalname: basename(filePath),
@@ -156,7 +143,7 @@ export class PluginFileManagerServer extends Plugin {
       });
     });
 
-    return getFileData({ app: this.app, file, storage: storageInstance, request: { body: {} } } as any);
+    return storageInstance.getFileData(file, {});
   }
 
   async loadStorages(options?: { transaction: any }) {
