+++ packages/plugins/@nocobase/plugin-file-manager/src/server/storages/index.ts
@@ -7,9 +7,10 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
-import { isURL } from '@nocobase/utils';
+import Path from 'path';
 import { StorageEngine } from 'multer';
 import urlJoin from 'url-join';
+import { isURL } from '@nocobase/utils';
 import { encodeURL, ensureUrlEncoded, getFileKey } from '../utils';
 
 export interface StorageModel {
@@ -46,7 +47,28 @@ export abstract class StorageType {
     return getFileKey(record);
   }
 
-  getFileData?(file: { [key: string]: any }): { [key: string]: any };
+  getFileData(file, meta = {}) {
+    const { [(this.constructor as typeof StorageType).filenameKey || 'filename']: name } = file;
+    // make compatible filename across cloud service (with path)
+    const filename = Path.basename(name);
+    const extname = Path.extname(filename);
+    const path = (this.storage.path || '').replace(/^\/|\/$/g, '');
+
+    const data = {
+      title: Buffer.from(file.originalname, 'latin1').toString('utf8').replace(extname, ''),
+      filename,
+      extname,
+      // TODO(feature): 暂时两者相同，后面 storage.path 模版化以后，这里只是 file 实际的 path
+      path,
+      size: file.size,
+      mimetype: file.mimetype,
+      meta,
+      storageId: this.storage.id,
+    };
+
+    return data;
+  }
+
   getFileURL(file: AttachmentModel, preview?: boolean): string | Promise<string> {
     // 兼容历史数据
     if (file.url && isURL(file.url)) {
